refactor: add barrel files for types
