+++ src/features/calendar/components/create-page.tsx
@@ -25,7 +25,7 @@ import {
 } from "../../../components/ui/select"
 import { createCalendarFromBulk } from "../services/calendar"
 import { getAllEmployees } from "@/features/employee/services/employee"
-import { Employee } from "@/features/employee/types/employee"
+import { Employee } from "@/features/employee/types"
 import { sortEmployeesByName } from "@/utils/employee"
 import { zodResolver } from "@hookform/resolvers/zod"
 import { ChevronDownIcon } from "lucide-react"
