+++ src/features/customer/components/overview-page.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { Customer } from "../types/customer"
+import { Customer } from "../types"
 import { Button } from "@/components/ui/button"
 import { DataTable, SortableColumn } from "@/components/ui/data-table"
 import {
