+++ src/features/customer/types/customer.ts
@@ -1,4 +1,4 @@
-import { Id } from "@/types/id"
+import { Id } from "@/types"
 
 export type Customer = {
   id?: Id
