+++ src/features/employee/components/details-form.tsx
@@ -9,7 +9,7 @@ import {
   FormLabel,
 } from "@/components/ui/form"
 import { Input } from "@/components/ui/input"
-import { Employee } from "@/features/employee/types/employee"
+import { Employee } from "@/features/employee/types"
 import { zodResolver } from "@hookform/resolvers/zod"
 import Link from "next/link"
 import { useForm } from "react-hook-form"
