+++ src/features/job/components/details-form.tsx
@@ -19,8 +19,8 @@ import {
 } from "@/components/ui/select"
 import { Textarea } from "@/components/ui/textarea"
 import { getAllCustomers } from "@/features/customer/services/customer"
-import { Customer } from "@/features/customer/types/customer"
-import { Job } from "@/features/job/types/job"
+import { Customer } from "@/features/customer/types"
+import { Job } from "@/features/job/types"
 import { zodResolver } from "@hookform/resolvers/zod"
 import Link from "next/link"
 import { useEffect, useState } from "react"
