+++ src/features/job/components/edit-page.tsx
@@ -5,8 +5,8 @@ import NotFound from "@/app/not-found"
 import LoadingPage from "@/components/skeleton/page"
 import { getCustomerById } from "@/features/customer/services/customer"
 import { editJob, getJobById } from "@/features/job/services/job"
-import { Job } from "@/features/job/types/job"
-import { Id } from "@/types/id"
+import { Job } from "@/features/job/types"
+import { Id } from "@/types"
 import { useRouter } from "next/navigation"
 import { useEffect, useState } from "react"
 import { toast } from "sonner"
